DAT-531 (P2 fix): split the read out so the pure module stays client-free
The P2 commit co-located readStageStaleness with the pure logic, so the unit
test transitively imported the postgres client (the at-import hang trap) and the
suite failed to load. Move the read to stage-staleness-read.ts; stage-staleness.ts
is pure again (RRstageStage logic only). 1217 tests green.
